Application Programming Interface (APIs)

Documenting the public user interface

Thermodynamics

MixedLayerModel.calc_LCL — Method

calc_LCL(u)

calculate the lifiting condensation level

MixedLayerModel.calc_qft0 — Method

calc_qft0(RHft, Gamma_q, sft0, Gamma_s)

calculate the initial free-tropospheric humidity given
ft RH, ft humidity lapse rate, initial ft dry static energy, 
and ft dry static energy lapse rate

MixedLayerModel.incloud_LWP — Method

incloud_LWP(u)

calulcate the in-cloud liquid water path

MixedLayerModel.moist_adiabat — Method

moist_adiabat(Tsurf, zft, p)
calculate moist adiabat given a surface temperature (Tsurf),
up to an altitude zft, with the parameters p
- first calculates the zLCL
- then calculates the moist adiabatic profile with dz=10m up to zft
returns (T,z) profile

MixedLayerModel.pres — Method

pres(z, T)

assumes hydrostatic balance

MixedLayerModel.q_l — Method

q_l(z, T, qt)

as difference between total specific humidity and saturation specific humidity
if undersaturated, then ql=0

MixedLayerModel.q_sat — Method

q_sat(z, T)

uses Clasius-Clapeyron relation with assumed constant 
latent heat of vaporization term L0=2.5e6

MixedLayerModel.rho — Method

rho(z, T)    

calculate density given altitude and temperature

MixedLayerModel.temp — Method

temp(z, h, qt)

uses saturation adjustment on the enthalpy

MixedLayerModel.temp_ft — Method

calculate actual moist adiabat by integrating
go up dry adiabat to LCL and then saturated adiabat

MixedLayerModel.Γs — Method

Γs - saturated adiabatic lapse rate

MixedLayerModel.ρref — Method

ρref(T)    

calculate reference density given temperature
and reference pressure pref

Surface fluxes

MixedLayerModel.H_0 — Method

define the surface enthalpy flux, H_surf
given prescribed sensible and latent heat fluxes

H_surf = (SHF + LHF) / ρref

MixedLayerModel.H_0 — Method

define surface enthalpy flux, H_surf
using bulk aerodynamic formula

H_surf = C * V * (h0 - h)

MixedLayerModel.Q_0 — Method

define the surface moisture flux, Q_surf
given prescribed latent heat flux

Q_surf = LHF / (Lv * ρref)

MixedLayerModel.Q_0 — Method

define surface moisture flux, Q_surf
using bulk aerodynamic formula

Q_surf = C * V * (q0 - q)

MixedLayerModel.calc_LHF — Method

calculate the latent heat flux

MixedLayerModel.calc_SHF — Method

calculate the sensible heat flux

Top fluxes

MixedLayerModel.H_zi — Method

H_zi(u, p, ent, LWP)

moist enthalpy flux into the mixed-layer from above at z=zi
H_zi = -we * (hft - hM)

MixedLayerModel.Q_zi — Method

Q_zi(u, p, ent, LWP)

moisture flux into the mixed-layer from above at z=zi
Q_zi = -we * (qft - qM)

MixedLayerModel.hjump — Method

hjump(u, p, LWP, p.fttype::co2dep)
defines the inversion jump for h 
    via linear regression to LES results

MixedLayerModel.hjump — Method

hjump(u, p, LWP, p.fttype::fixedFT)
defines h+(z) in free troposphere -- given Gamma_s and Gamma_q

MixedLayerModel.hjump — Method

hjump(u, p, LWP, p.fttype::twocol)

MixedLayerModel.qjump — Method

qjump(u, p, LWP, p.fttype::co2dep)
defines the inversion jump for qt 
    via linear regression to LES results

MixedLayerModel.qjump — Method

qjump(u, p, LWP, p.fttype::fixedFT)
defines qt+(z) in free troposphere -- given Gamma_q
minimum value of qft of 2 g/kg

MixedLayerModel.qjump — Method

qjump(u, p, LWP, p.fttype::twocol)

specific humidity above cloud given fixed RH=0.2
and saturation calculated at Tft and fixed 1500 m

Radiation

MixedLayerModel.calc_cloudtop_RAD — Method

returns the prescribed cloud-top radiative cooling ΔR

MixedLayerModel.calc_cloudtop_RAD — Method

calculate the net ΔR at cloud-top based on CO2
balance between upwelling and downwelling longwave
downwelling longwave is based on an effective temperature
which is empirically fit to LES

gives ΔR ≈ 80 W/m2 for 400 ppm CO2

MixedLayerModel.calc_surf_RAD — Method

calculate net SW and LW radiation at the surface

MixedLayerModel.trop_sst — Method

trop_sst(u, p, LWP)
tropical SST specified as a deviation from a base state p.Ts400
with warming from export from the subtropics (proportional to all-sky albedo)
and warming directly from GHG that depends on the ECS parameter

Entrainment

MixedLayerModel.sv_jump — Method

sv_jump(u, p, LWP)

calculate the jump in virtual liquid static energy (sv)
across the inversion
Δsv = Δs + CpΔTv - CpΔT
Tv = (Rd*(1-qt) + Rv*qv)/Rd * T

MixedLayerModel.we — Method

we(u, p, zb, LWP, etype::bflux)

entrainment velocity based on buoyancy flux
without radiation

integral is calculated analytically

MixedLayerModel.we — Method

we(u, p, zb, LWP, etype::enBal)

entrainment velocity obtained via energy balance requirement
w = ΔR / (Δsv * ρref)

MixedLayerModel.we — Method

we(u, p, zb, LWP, etype::fixed)

fixed entrainment velocity of 7 mm/s

MLM ODE

MixedLayerModel.mlm — Method

mlm(du, u, p, t)    

define the coupled ODE
  dzi/dt = we - D*zi
  dhM/dt = -dE/dz = 1/zi * (Hzi - H0 + dR/rho)
  dqM/dt = -dW/dz = 1/zi * (Qzi - Q0)
  dSST/dt = 1/c * (SWnet - LWnet - SHF - LHF - OHU)

Diagnostics

MixedLayerModel.calc_OHU — Method

calc_OHU(u, p, p.stype::fixSST)

u is the state vector [zi, hM, qM, SST, CF]
p is the parameter object

calculates the ocean heat uptake as the residual 
between the radiative fluxes and the LHF + SHF

MixedLayerModel.calc_OHU — Method

calc_OHU(u, p, p.stype::varSST)

u is the state vector [zi, hM, qM, SST, CF]
p is the parameter object

returns p.OHU

MixedLayerModel.calc_bflux — Method

calc_bflux(u, p, zarr, etype::bflux)

u is the state vector [zi, hM, qM, SST, CF]
p is the parameter object
zarr is an array of altitudes from 0 to some maxz > zi

calculates the buoyancy flux for plotting