Application Programming Interface (APIs)

Documenting the public user interface

Thermodynamics

MixedLayerModel.RH — Method

RH(z, h, qt)

relative humidity is the ratio of 
total specific humidity to saturation
max value is 1

MixedLayerModel.calc_LCL — Method

calc_LCL(zi, hM, qtM)

calculate the lifiting condensation level

MixedLayerModel.calc_LWP — Method

calc_LWP(zi, hM, qtM)

calulcate the liquid water path

MixedLayerModel.calc_qft0 — Method

calc_qft0(RHft, Gamma_q, sft0, Gamma_s)

calculate the initial free-tropospheric humidity given
ft RH, ft humidity lapse rate, initial ft dry static energy, 
and ft dry static energy lapse rate

MixedLayerModel.pres — Method

pres(z, T)

assumes hydrostatic balance

MixedLayerModel.q_l — Method

q_l(z, T, qt)

as difference between total specific humidity and saturation specific humidity
if undersaturated, then ql=0

MixedLayerModel.q_sat — Method

q_sat(z, T)

uses Clasius-Clapeyron relation with assumed constant 
latent heat of vaporization term L0=2.5e6

MixedLayerModel.q_v — Method

q_v(z, T, qt)

as the minimum between the total specific humidty and saturation specific humidity

MixedLayerModel.rho — Method

rho(z, T)    

calculate density given altitude and temperature

MixedLayerModel.temp — Method

temp(z, h, qt)

uses saturation adjustment on the enthalpy
if no zero is found, then set temp = 0°C

MixedLayerModel.theta — Method

theta(z,h,qt)

calculate the potential temperature

MixedLayerModel.ρref — Method

ρref(T)    

calculate reference density given temperature
and reference pressure pref

Surface fluxes

MixedLayerModel.H_0 — Method

define the surface enthalpy flux, H_surf
given prescribed sensible and latent heat fluxes

H_surf = (SHF + LHF) / ρref

MixedLayerModel.H_0 — Method

define surface enthalpy flux, H_surf
using bulk aerodynamic formula

H_surf = C * V * (h0 - h+)

MixedLayerModel.Q_0 — Method

define the surface moisture flux, Q_surf
given prescribed latent heat flux

Q_surf = LHF / (Lv * ρref)

MixedLayerModel.Q_0 — Method

define surface moisture flux, Q_surf
using bulk aerodynamic formula

Q_surf = C * V * (q0 - q+)

MixedLayerModel.calc_LHF — Method

calculate the latent heat flux

MixedLayerModel.calc_SHF — Method

calculate the sensible heat flux

Top fluxes

MixedLayerModel.H_zi — Method

H_zi(u, p)

enthalpy flux into the mixed-layer from above at z=zi
H_zi = -we * (hft - hM)

MixedLayerModel.Q_zi — Method

Q_zi(u, p)

moisture flux into the mixed-layer from above at z=zi
Q_zi = -we * (qft - qM)

MixedLayerModel.h_ft — Method

h_ft(z, p)

defines h+(z) in free troposphere -- given Gamma_s and Gamma_q

MixedLayerModel.q_ft — Method

q_ft(z, p)

defines qt+(z) in free troposphere -- given Gamma_q

Radiation

MixedLayerModel.calc_cloudtop_RAD — Method

calculate the net OLR at cloud-top based on CO2
this is the 3-layer atmosphere model

MixedLayerModel.calc_cloudtop_RAD — Method

calculate the net OLR at cloud-top based on CO2
this is the 3-layer atmosphere model

MixedLayerModel.calc_surf_RAD — Method

calculate net SW and LW radiation at the surface

Entrainment

MixedLayerModel.we — Method

we(u, p, etype::bflux)

entrainment velocity based on buoyancy flux
without radiation

integral is calculated analytically

MixedLayerModel.we — Method

we(u, p, etype::enBal)

entrainment velocity obtained via energy balance requirement
w = ΔR / (Δs_vli * ρref)

MixedLayerModel.we — Method

we(u, p, etype::fixed)

fixed entrainment velocity of 1.5 mm/s

MLM ODE

MixedLayerModel.mlm — Method

mlm(du, u, p, t)    

define the coupled ODE
  dzi/dt = D*zi - we
  dhM/dt = -dE/dz = 1/zi * (Hzi - H0 + dR/rho)
  dqM/dt = -dW/dz = 1/zi * (Qzi - Q0)
  dSST/dt = 1/c * (SWnet - LWnet - SHF - LHF - OHU)

MLM solve

MixedLayerModel.run_mlm — Method

run_mlm(params, filename="default.txt")

run MLM simulation with given parameters
and save output to file

MixedLayerModel.run_mlm_from_init — Function

run_mlm_from_init(u0, params, filename="default.txt")

run MLM simulation with given parameters
from initial state u0
and save output to file

Diagnostics

MixedLayerModel.calc_OHU — Method

calc_OHU(u, p, p.stype::fixSST)

u is the state vector [zi, hM, qM, SST]
p is the parameter object

calculates the ocean heat uptake as the residual 
between the radiative fluxes and the LHF + SHF

MixedLayerModel.calc_OHU — Method

calc_OHU(u, p, p.stype::varSST)

u is the state vector [zi, hM, qM, SST]
p is the parameter object

returns p.OHU

MixedLayerModel.calc_bflux — Method

calc_bflux(u, p, zarr, etype::bflux)

u is the state vector [zi, hM, qM, SST]
p is the parameter object
zarr is an array of altitudes from 0 to some maxz > zi

calculates the buoyancy flux for plotting