Application Programming Interface (APIs)

Documenting the public user interface

Thermodynamics

MixedLayerModel.calc_LCL — Method

calc_LCL(u)

calculate the lifiting condensation level

MixedLayerModel.incloud_LWP — Method

incloud_LWP(u)

calulcate the in-cloud liquid water path

MixedLayerModel.moist_adiabat — Method

moist_adiabat(Tsurf, zft, p)
calculate moist adiabat given a surface temperature (Tsurf),
up to an altitude zft, with the parameters p
- first calculates the zLCL
- then calculates the moist adiabatic profile with dz=10m up to zft
returns (T,z) profile

MixedLayerModel.pres — Method

pres(z, T)

assumes hydrostatic balance

MixedLayerModel.q_l — Method

q_l(z, T, qt)

as difference between total specific humidity and saturation specific humidity
if undersaturated, then ql=0

MixedLayerModel.q_sat — Method

q_sat(z, T)

uses Clasius-Clapeyron relation with assumed constant 
latent heat of vaporization term L0=2.5e6

MixedLayerModel.rho — Method

rho(z, T)    

calculate density given altitude and temperature

MixedLayerModel.temp — Method

temp(z, s, qt)

uses saturation adjustment on the liquid water static energy (s)

MixedLayerModel.temp_ft — Method

calculate actual moist adiabat by integrating
go up dry adiabat to LCL and then saturated adiabat

MixedLayerModel.Γs — Method

Γs - saturated adiabatic lapse rate

MixedLayerModel.ρref — Method

ρref(T)    

calculate reference density given temperature
and reference pressure pref

Surface fluxes

MixedLayerModel.Q_0 — Method

define the surface moisture flux, Q_surf
given prescribed latent heat flux

Q_surf = LHF / (Lv * ρref)

MixedLayerModel.Q_0 — Method

define surface moisture flux, Q_surf
using bulk aerodynamic formula

Q_surf = C * V * (q0 - q)

MixedLayerModel.S_0 — Method

define the surface liquid static energy flux, S_surf
given prescribed sensible and latent heat fluxes

S_surf = SHF / ρref

MixedLayerModel.S_0 — Method

define surface liquid static energy flux, S_surf
using bulk aerodynamic formula

S_surf = C * V * (s0 - s)

MixedLayerModel.calc_LHF — Method

calculate the latent heat flux

MixedLayerModel.calc_SHF — Method

calculate the sensible heat flux

Top fluxes

MixedLayerModel.Q_zi — Method

Q_zi(u, p, ent, LWP)

moisture flux into the mixed-layer from above at z=zi
Q_zi = -we * (qft - qM)

MixedLayerModel.S_zi — Method

S_zi(u, p, ent, LWP)

energy flux into the mixed-layer from above at z=zi
S_zi = -we * (sft - sM)

MixedLayerModel.qjump — Method

qjump(u, p, LWP, fttype::Union{twocol, fixEIS, fixedFT, co2EIS})

specific humidity above cloud given fixed RHft
and saturation calculated at Tft

MixedLayerModel.sjump — Method

sjump(u, p, LWP, p.fttype::co2EIS)
defines s+(z) in free troposphere given EIS and dTdz

MixedLayerModel.sjump — Method

sjump(u, p, LWP, p.fttype::fixEIS)
defines s+(z) in free troposphere given EIS and dTdz

MixedLayerModel.sjump — Method

sjump(u, p, LWP, p.fttype::fixedFT)
defines s+(z) in free troposphere -- given Gamma_s and Gamma_q

Radiation

MixedLayerModel.calc_cloudtop_RAD — Method

calc_cloudtop_RAD(u, p, LWP, rtype::fixRad)

returns the prescribed cloud-top radiative cooling ΔR

MixedLayerModel.calc_cloudtop_RAD — Method

calc_cloudtop_RAD(u, p, LWP, rtype::varRad)

calculate the net ΔR at cloud-top based on CO2
balance between upwelling and downwelling longwave
downwelling longwave is based on an effective temperature
which is empirically fit to LES

gives ΔR ≈ 75 W/m2 for 400 ppm CO2

MixedLayerModel.calc_surf_RAD — Method

calc_surf_RAD(u, p, LWP)

calculate net SW and LW radiation at the surface

MixedLayerModel.cloud_albedo — Method

cloud_albedo(LWP)

albedo of the cloud given LWP in kg/m^2
fit from LES experiments

MixedLayerModel.cloud_emissivity — Method

cloud_emissivity(LWP)

emissivity of the cloud as a function of LWP in kg/m2
ϵ = 1 - exp(-a0 * LWP) with a0 = 0.15 m^2/g

based on Stephens 1978 part II: eq 15 and 16

MixedLayerModel.ΔTa — Method

ΔTa(u, p, LWP, p.wvtype::wvON)
writes the difference between cloud-top temperature and downwelling emission
temperature as a function of CO2 and H2O above-cloud

MixedLayerModel.ΔTa — Method

ΔTa(u, p, LWP, p.wvtype::wvRADOFF)
writes the difference between cloud-top temperature and downwelling emission
temperature as a function of CO2 
H2O above-cloud is FIXED by p.qft_rad

Entrainment

MixedLayerModel.sv_jump — Method

sv_jump(u, p, LWP)

Δsv = Δs + CpΔTv - CpΔT
    = Δs + Cp(Rv/Rd - 1)(Tft*qft - T_zi*qM) + Cp(Rv/Rd)(T_zi*ql_zi)

Tv = (Rd*(1-qt) + Rv*qv)/Rd * T

jump in virtual liquid static energy across inversion
proportional to buoyancy jump
used in energy balance entrainment

MixedLayerModel.we — Method

we(u, p, zb, LWP, etype::bflux)

entrainment velocity based on buoyancy flux
without radiation

integral is calculated analytically

MixedLayerModel.we — Method

we(u, p, zb, LWP, etype::enBal)

entrainment velocity obtained via energy balance requirement
w = ΔR / (Δsv * ρref)
Δsv = the jump in virtual liquid static energy
sv = Cp*Tv + g*z - Lv*ql = s + Cp(Tv - T)

Cloud fraction

MixedLayerModel.calc_decoupling — Method

calc_decoupling(u, p)

calculates the stability parameter, s
also called the minimal decloupling parameter
in Bretherton and Wyant (1997)
De = (LHF / ΔR) * (zc / zi)
where zc is the cloud thickness (zi-zb)

MixedLayerModel.cloud_fraction — Method

cloud_fraction(u, p, zb, LWP)

calculates the cloud fraction 
inspired by Chung et al 2012
uses a logistic function to smoothly interpolate between 
100% and 20% cloud fraction based on the value of the
stability parameter

MLM ODE

MixedLayerModel.mlm — Method

mlm(du, u, p, t)    

define the coupled ODE
  dzi/dt = we - D*zi
  dsM/dt = -dE/dz = -1/zi * (Szi - S0 + dR/rho)
  dqM/dt = -dW/dz = -1/zi * (Qzi - Q0)
  dSST/dt = (SWnet - LWnet - SHF - LHF - OHU) / τ_SST
  dCF/dt = (CF' - CF) / τ_CF

Diagnostics

MixedLayerModel.calc_OHU — Method

calc_OHU(u, p, p.stype::fixSST)

u is the state vector [zi, sM, qM, SST, CF]
p is the parameter object

calculates the ocean heat uptake as the residual 
between the radiative fluxes and the LHF + SHF

MixedLayerModel.calc_OHU — Method

calc_OHU(u, p, p.stype::varSST)

u is the state vector [zi, sM, qM, SST, CF]
p is the parameter object

returns p.OHU

MixedLayerModel.calc_bflux — Method

calc_bflux(u, p, zarr, etype::bflux)

u is the state vector [zi, sM, qM, SST, CF]
p is the parameter object
zarr is an array of altitudes from 0 to some maxz > zi

calculates the buoyancy flux for plotting